Create JRE-Download2.bat
Starting with podcast # 2038 Trae Tha Truth, and interspersed with subsequent podcasts, Spotify appears to have changed the "folder" numbers that contain the various qualities. If you are getting errors on the original JRE Download .bat file, try this one and see if it remedies it. I'll be updating the main guide with this new information.